Moved threadpool benchmarks to a seperate file.
